Fix snow shader see-through-shadows
Bug was caused by renderer and mesh not overlapping. Reason for this was having snow only is an offset of
landscape - since snow is separate with an unused collider, renderer and mesh moved back together and
collider swapped to box (this also simplified the raycast a lot).
TODO: change raycast to simple 2D box collision.